[refactor] Renaming `config/internal` package to `config/common`
