+++ pkg/config/authorization.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/3scale-labs/authorino/pkg/config/authorization"
-	"github.com/3scale-labs/authorino/pkg/config/internal"
+	"github.com/3scale-labs/authorino/pkg/config/common"
 )
 
 type AuthorizationConfig struct {
@@ -12,7 +12,7 @@ type AuthorizationConfig struct {
 	JWT *authorization.JWTClaims `yaml:"jwt"`
 }
 
-func (self *AuthorizationConfig) Call(ctx internal.AuthContext) (bool, error) {
+func (self *AuthorizationConfig) Call(ctx common.AuthContext) (bool, error) {
 	switch {
 	case self.OPA != nil:
 		return self.OPA.Call(ctx)
