+++ pkg/config/identity/mtls.go
@@ -1,13 +1,13 @@
 package identity
 
 import (
-	"github.com/3scale-labs/authorino/pkg/config/internal"
+	"github.com/3scale-labs/authorino/pkg/config/common"
 )
 
 type MTLS struct {
 	PEM string `yaml:"pem"`
 }
 
-func (self *MTLS) Call(ctx internal.AuthContext) (interface{}, error) {
+func (self *MTLS) Call(ctx common.AuthContext) (interface{}, error) {
 	return "Authenticated with mTLS", nil // TODO: implement
 }
